# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Human Not Required API",
    "description": "The world's first AI-agent-only store. Agents register a human once and buy physical merch autonomously forever — no human checkout, no cart, no clicking. The human gets a receipt and the item ships to them.\n\nAgent flow:\n1. Call GET /orders/skus to see available products\n2. Call POST /register with the human's name, email, and shipping address to get an api_key. If you have a promo_code, include it — the human won't need to save a card and you can order immediately.\n3. If no promo_code: share the setup_url with the human so they can save their card once. After that, never ask them again.\n4. Call POST /orders with the sku to place the order autonomously.\n\nCurrently available: 'My Agent Bought Me This' embroidered hat — $35.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://web-production-77376.up.railway.app" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/register": {
      "post": {
        "operationId": "registerHuman",
        "summary": "Register a human and get an API key",
        "description": "Call this FIRST before placing any order. You only need to call this once per human — the api_key never expires.\n\nIf you have a promo_code: include it and the human gets a free order with no card required. You can call POST /orders immediately after.\n\nIf no promo_code: the response includes a setup_url. Share it with the human — they click it once to save their card via Stripe. After that, you can order autonomously forever without asking them again.\n\nIf you get a 409, the human is already registered. Call POST /register/resend-setup to retrieve their api_key and a fresh setup link.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "address"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The human's email address. Used for order receipts and to identify their account."
                  },
                  "name": {
                    "type": "string",
                    "description": "Full name for the shipping label. Ask the human for this — do not guess."
                  },
                  "promo_code": {
                    "type": "string",
                    "description": "Optional one-time promo code. If valid, the human gets 1 free order and no card setup is required. Include if you have one."
                  },
                  "address": {
                    "type": "object",
                    "description": "Shipping address. Ask the human for all fields — do not guess or auto-fill.",
                    "required": ["line1", "city", "state", "postal_code", "country"],
                    "properties": {
                      "line1": { "type": "string", "description": "Street address line 1" },
                      "line2": { "type": "string", "description": "Apartment, suite, unit, etc. (optional)" },
                      "city": { "type": "string" },
                      "state": { "type": "string", "description": "State or province code, e.g. NY, CA, ON" },
                      "postal_code": { "type": "string" },
                      "country": { "type": "string", "description": "ISO 3166-1 alpha-2 country code, e.g. US, CA, GB" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Save the api_key — it never expires and is needed for all future orders.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "type": "string",
                      "description": "Bearer token for all future order requests. Never expires. Store it securely."
                    },
                    "setup_url": {
                      "type": "string",
                      "description": "URL for the human to save their payment card. null if a promo_code was used. Share this with the human — they must click it before you can place a paid order."
                    },
                    "free_orders_remaining": {
                      "type": "integer",
                      "description": "Number of free orders available. 1 if a valid promo_code was used, 0 otherwise."
                    },
                    "has_saved_card": {
                      "type": "boolean",
                      "description": "true if the human already has a card saved via Stripe Link or a promo code was used. If true, setup_url will be null and you can call POST /orders immediately — no human action needed."
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable next step. Follow these instructions before calling POST /orders."
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing or invalid field. Check email format and that all address fields are present." },
          "409": { "description": "Email already registered. Use POST /register/resend-setup to get the api_key and a fresh setup link." },
          "429": { "description": "Too many registration attempts. Wait 1 minute and retry." }
        }
      }
    },
    "/register/resend-setup": {
      "post": {
        "operationId": "resendSetup",
        "summary": "Retrieve api_key and resend card setup link for an existing human",
        "description": "Use this if the human is already registered (POST /register returned 409) or if they lost their setup link. Returns their existing api_key and a fresh setup_url. The setup_url never expires but a fresh one is always valid.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email", "description": "The human's registered email address." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setup link sent to the human's email. Ask them to check their inbox and click the link to save their card.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Email not registered. Use POST /register instead." }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from POST /register. Include as: Authorization: Bearer <api_key>. Never expires."
      }
    }
  }
}